Explicitly set last modified date (#23)
For some reason, the logic introduced in jekyll/jekyll-sitemap@7f7f989
as part of resolving jekyll/jekyll-sitemap#62 does not work as expected.
A <lastmod> element is generated in any case, even though the
jekyll-last-modified-at plugin is not in play, and (as seen in previous
commit), no date: front-matter is present.

Ideally, it would be possible to eliminate <lastmod> altogether, but
since that appears to be an option, this commit explicitly sets the
last_modified_at: variable to the date of the last substantive change to
the document (which was 053f397043b3b6b1ecb061bf8a02ea3d6a9947e8).